Fixes the regex replacing one of the version numbers in the pluginMainFile.
Before, <%= pluginVersionConstant %> was literally matched by the regex, which is incorrect.
Now, it's actually replaced by the pluginVersionConstant before the regex matches.
In order to do so, the regex needed to be created using `new RegExp()`.
Regexes created with // can't have dynamic content.
Because we're now dealing with two types of regex formats in the config (strings and regexes with //),
the task needed to be modified as well. This change also secure backwards compatibility
with other plugins using this repo.