Watch: Remove default JS and CSS watchers
As we don't have to have JS or CSS in any of the plugins, having this configuration requires each plugin to overwrite or work-around these configs if they don't.

Having an empty configuration value (for files.sass) will break the watch command and have it stuck in an unending loop.

This configuration should be added in each plugin anyway, to have a better understanding of which files are being watched and what checks are performed when they are being changed.

Note that PHP is missing from this configuration to begin with.